requirements.txt
ultralytics>=8.0.0
opencv-python
pyrealsense2
Pillow
numpy